+++ babel.config.js
@@ -2,6 +2,7 @@ module.exports = {
   presets: [
     ['@babel/preset-env', { targets: { node: 'current' } }],
     '@babel/preset-typescript',
+    '@babel/preset-react',
     [
       'next/babel',
       {
