+++ pages/submission/dcc/dashboard.tsx
@@ -17,7 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { createPage } from 'global/utils/pages';
+import createPage from 'global/utils/pages/createPage';
 import { isDccMember } from 'global/utils/egoJwt';
 import DccDashboard from 'components/pages/submission-system/dcc';
