+++ pages/submission/program/[shortName]/clinical-submission.tsx
@@ -19,7 +19,7 @@
 
 import React from 'react';
 
-import { createPage } from 'global/utils/pages';
+import createPage from 'global/utils/pages/createPage';
 import ProgramClinicalSubmission from 'components/pages/submission-system/program-clinical-submission';
 import { canReadProgram, canWriteProgramData } from 'global/utils/egoJwt';
 import { useProgramCheckEffect } from 'global/hooks/useProgramCheckEffect';
