+++ pages/submission/program/index.tsx
@@ -20,7 +20,7 @@
 import React from 'react';
 
 import { isDccMember } from 'global/utils/egoJwt';
-import { createPage } from 'global/utils/pages';
+import createPage from 'global/utils/pages/createPage';
 import ProgramsPage from 'components/pages/submission-system/programs';
 import SIDE_MENU_PROGRAM_LIST from 'components/pages/submission-system/SIDE_MENU_PROGRAM_LIST.gql';
 import PROGRAMS_LIST_QUERY from 'components/pages/submission-system/programs/PROGRAMS_LIST_QUERY.gql';
