+++ pages/user/index.tsx
@@ -20,12 +20,12 @@
 import React from 'react';
 import get from 'lodash/get';
 
-import { createPage } from 'global/utils/pages';
+import createPage from 'global/utils/pages/createPage';
 import { decodeToken } from 'global/utils/egoJwt';
 import { UserPage } from 'components/pages/user';
 
 export default createPage({
-  getInitialProps: async context => {
+  getInitialProps: async (context) => {
     const { egoJwt, asPath, query, res } = context;
     const data = decodeToken(egoJwt);
     const firstName = get(data, 'context.user.firstName', '');
